pyproject.toml
[project]
name = "abb-icp"
version = "0.1.0"
description = "ICP analysis example code."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "jupyterlab>=4.4.3",
    "matplotlib>=3.10.3",
    "openai>=1.91.0",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.0",
    "tabulate>=0.9.0",
]
